+++ constant/enums/trading_enums/trading.go
@@ -54,10 +54,11 @@ const (
 
 // 交易所枚举
 const (
-	ExchangeBinance Exchange = "Binance" // 币安
-	ExchangeOKX     Exchange = "OKX"     // OKX
-	ExchangeLTP     Exchange = "LTP"     // LTP
-	ExchangeDeribit Exchange = "Deribit" // Deribit
+	ExchangeBinance     Exchange = "Binance"     // 币安
+	ExchangeOKX         Exchange = "OKX"         // OKX
+	ExchangeLTP         Exchange = "LTP"         // LTP
+	ExchangeDeribit     Exchange = "Deribit"     // Deribit
+	ExchangeHyperliquid Exchange = "Hyperliquid" // Hyperliquid
 )
 
 // 币对品种枚举（与市场类型对应）